requirements.txt
gevent
pyzmq
